protecting routes

+++ controllers/contactController.js
@@ -5,15 +5,15 @@ const Contact = require("../models/contactModel");
 
 // @desc Get all contacts
 // @route GET /api/contacts
-// @access public
+// @access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id }); // For the user who is logged in
     res.status(200).json(contacts); 
 });
 
 // @desc Create new contacts
 // @route POST /api/contacts
-// @access public
+// @access private
 const createContact = asyncHandler(async (req, res) => {
     console.log("The request body is : ", req.body);
     const {name, email, phone } = req.body;
@@ -25,13 +25,14 @@ const createContact = asyncHandler(async (req, res) => {
         name, 
         email,
         phone,
+        user_id: req.user.id,
     });
     res.status(201).json(contact);
 });
 
 // @desc Get a contact
 // @route GET /api/contacts/:id
-// @access public
+// @access private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
@@ -43,12 +44,17 @@ const getContact = asyncHandler(async (req, res) => {
 
 // @desc Update a contact
 // @route PUT /api/contacts/:id
-// @access public
+// @access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User does not have permission to update other user contacts");
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -61,12 +67,17 @@ const updateContact = asyncHandler(async (req, res) => {
 
 // @desc Delete a contact
 // @route DELETE /api/contacts/:id
-// @access public
+// @access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User does not have permission to delete other user contacts");
     }
 
     await Contact.deleteOne({_id: req.params.id});
